refactor(apollo_l1_provider): move contentbuilders to unit tests
- These should only be used in unit tests anyway, no point exporting
  them in test_utils.
- About to add private fields to the L1Provider which require the
  test_utils to be a sub-module of the L1Provider, which it isn't at the
  moment --- it could be, but see point 1 :)
