+++ crates/apollo_l1_provider/src/l1_provider_tests.rs
@@ -1,4 +1,5 @@
 use std::collections::HashSet;
+use std::mem;
 use std::sync::Arc;
 use std::time::Duration;
 
@@ -16,13 +17,188 @@ use itertools::Itertools;
 use pretty_assertions::assert_eq;
 use rstest::rstest;
 use starknet_api::block::BlockNumber;
+use starknet_api::executable_transaction::L1HandlerTransaction;
 use starknet_api::test_utils::l1_handler::{executable_l1_handler_tx, L1HandlerTxArgs};
 use starknet_api::transaction::TransactionHash;
 use starknet_api::tx_hash;
 
 use crate::bootstrapper::{Bootstrapper, CommitBlockBacklog, SyncTaskHandle};
-use crate::test_utils::{l1_handler, FakeL1ProviderClient, L1ProviderContentBuilder};
+use crate::l1_provider::L1Provider;
+use crate::soft_delete_index_map::SoftDeleteIndexMap;
+use crate::test_utils::{l1_handler, FakeL1ProviderClient};
+use crate::transaction_manager::TransactionManager;
 use crate::ProviderState;
+
+// Represents the internal content of the L1 provider for testing.
+// Enables customized (and potentially inconsistent) creation for unit testing.
+#[derive(Default)]
+pub struct L1ProviderContent {
+    tx_manager_content: Option<TransactionManagerContent>,
+    state: Option<ProviderState>,
+    current_height: Option<BlockNumber>,
+}
+
+impl L1ProviderContent {
+    #[track_caller]
+    pub fn assert_eq(&self, l1_provider: &L1Provider) {
+        if let Some(tx_manager_content) = &self.tx_manager_content {
+            tx_manager_content.assert_eq(&l1_provider.tx_manager);
+        }
+
+        if let Some(state) = &self.state {
+            assert_eq!(&l1_provider.state, state);
+        }
+
+        if let Some(current_height) = &self.current_height {
+            assert_eq!(&l1_provider.current_height, current_height);
+        }
+    }
+}
+
+impl From<L1ProviderContent> for L1Provider {
+    fn from(content: L1ProviderContent) -> L1Provider {
+        L1Provider {
+            tx_manager: content.tx_manager_content.map(Into::into).unwrap_or_default(),
+            // Defaulting to Pending state, since a provider with a "default" Bootstrapper
+            // is functionally equivalent to Pending for testing purposes.
+            state: content.state.unwrap_or(ProviderState::Pending),
+            current_height: content.current_height.unwrap_or_default(),
+        }
+    }
+}
+
+#[derive(Debug, Default)]
+pub struct L1ProviderContentBuilder {
+    tx_manager_content_builder: TransactionManagerContentBuilder,
+    state: Option<ProviderState>,
+    current_height: Option<BlockNumber>,
+}
+
+impl L1ProviderContentBuilder {
+    pub fn new() -> Self {
+        Self::default()
+    }
+
+    pub fn with_state(mut self, state: ProviderState) -> Self {
+        self.state = Some(state);
+        self
+    }
+
+    pub fn with_txs(mut self, txs: impl IntoIterator<Item = L1HandlerTransaction>) -> Self {
+        self.tx_manager_content_builder = self.tx_manager_content_builder.with_txs(txs);
+        self
+    }
+
+    pub fn with_rejected(mut self, txs: impl IntoIterator<Item = L1HandlerTransaction>) -> Self {
+        self.tx_manager_content_builder = self.tx_manager_content_builder.with_rejected(txs);
+        self
+    }
+
+    pub fn with_height(mut self, height: BlockNumber) -> Self {
+        self.current_height = Some(height);
+        self
+    }
+
+    pub fn with_committed(mut self, tx_hashes: impl IntoIterator<Item = TransactionHash>) -> Self {
+        self.tx_manager_content_builder = self.tx_manager_content_builder.with_committed(tx_hashes);
+        self
+    }
+
+    pub fn build(self) -> L1ProviderContent {
+        L1ProviderContent {
+            tx_manager_content: self.tx_manager_content_builder.build(),
+            state: self.state,
+            current_height: self.current_height,
+        }
+    }
+
+    pub fn build_into_l1_provider(self) -> L1Provider {
+        self.build().into()
+    }
+}
+
+// Represents the internal content of the TransactionManager.
+// Enables customized (and potentially inconsistent) creation for unit testing.
+#[derive(Debug, Default)]
+struct TransactionManagerContent {
+    pub uncommitted: Option<Vec<L1HandlerTransaction>>,
+    pub rejected: Option<Vec<L1HandlerTransaction>>,
+    pub committed: Option<HashSet<TransactionHash>>,
+}
+
+impl TransactionManagerContent {
+    #[track_caller]
+    fn assert_eq(&self, tx_manager: &TransactionManager) {
+        if let Some(uncommitted) = &self.uncommitted {
+            assert_eq!(
+                uncommitted,
+                &tx_manager.uncommitted.txs.values().map(|tx| tx.transaction.clone()).collect_vec()
+            );
+        }
+
+        if let Some(committed) = &self.committed {
+            assert_eq!(committed, &tx_manager.committed);
+        }
+
+        if let Some(rejected) = &self.rejected {
+            assert_eq!(
+                rejected,
+                &tx_manager.rejected.txs.values().map(|tx| tx.transaction.clone()).collect_vec()
+            );
+        }
+    }
+}
+
+impl From<TransactionManagerContent> for TransactionManager {
+    fn from(mut content: TransactionManagerContent) -> TransactionManager {
+        let txs: Vec<_> = mem::take(&mut content.uncommitted).unwrap_or_default();
+        TransactionManager {
+            uncommitted: SoftDeleteIndexMap::from(txs),
+            committed: content.committed.unwrap_or_default(),
+            rejected: SoftDeleteIndexMap::default(),
+        }
+    }
+}
+
+#[derive(Debug, Default)]
+struct TransactionManagerContentBuilder {
+    uncommitted: Option<Vec<L1HandlerTransaction>>,
+    rejected: Option<Vec<L1HandlerTransaction>>,
+    committed: Option<HashSet<TransactionHash>>,
+}
+
+impl TransactionManagerContentBuilder {
+    fn with_txs(mut self, txs: impl IntoIterator<Item = L1HandlerTransaction>) -> Self {
+        self.uncommitted = Some(txs.into_iter().collect());
+        self
+    }
+
+    fn with_rejected(mut self, rejected: impl IntoIterator<Item = L1HandlerTransaction>) -> Self {
+        self.rejected = Some(rejected.into_iter().collect());
+        self
+    }
+
+    fn with_committed(mut self, tx_hashes: impl IntoIterator<Item = TransactionHash>) -> Self {
+        self.committed = Some(tx_hashes.into_iter().collect());
+        self
+    }
+
+    fn build(self) -> Option<TransactionManagerContent> {
+        if self.is_default() {
+            return None;
+        }
+
+        Some(TransactionManagerContent {
+            uncommitted: self.uncommitted,
+            committed: self.committed,
+            rejected: self.rejected,
+        })
+    }
+
+    fn is_default(&self) -> bool {
+        self.uncommitted.is_none() && self.committed.is_none()
+    }
+}
 
 fn setup_rejected_transactions() -> super::L1Provider {
     let tx_id1 = 1;
